Added code for SporepediaLoadListener
Now it will replace some of the message boxes that come up when attempting to load an adventure with missing creations. However, this code is still unfinished as I need to figure out how to fetch the resource key of the adventure that was attempted to be loaded. And maybe wait for ``Simulator::ScenarioModeUI`` class to be implemented in the SDK so I can find out how that class works in the first place.